Extract getExecOptions method to common
